requirements.txt
pandas
sshtunnel
python-dotenv
openpyxl
PyMySQL
python-dotenv
openpyxl
PyMySQL
